requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
shioaji==1.3.0
google-generativeai>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv==1.0.1
aiosqlite==0.20.0
apscheduler==3.10.4
httpx==0.27.0
jinja2==3.1.4
